Refactor and style fixes
